feat: jwt를 이용한 회원용 API로 변경

+++ RisingCamp-Spring-Template/src/main/java/com/example/demo/src/controller/UserController.java
@@ -10,6 +10,7 @@ import com.example.demo.src.domain.dto.address.PostAddressReq;
 import com.example.demo.src.domain.dto.sign.PostLoginReq;
 import com.example.demo.src.domain.dto.user.GetUserRes;
 import com.example.demo.src.domain.dto.user.PostUserReq;
+import com.example.demo.src.domain.dto.user.PostUserRes;
 import com.example.demo.src.domain.entitiy.user.Address;
 import com.example.demo.src.domain.entitiy.user.PushNotificationAgreement;
 import com.example.demo.src.domain.entitiy.user.User;
@@ -17,6 +18,7 @@ import com.example.demo.src.service.UserProvider;
 import com.example.demo.src.service.UserService;
 import com.example.demo.utils.JwtService;
 import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.web.bind.annotation.*;
@@ -31,28 +33,17 @@ import static com.example.demo.utils.ValidationRegex.isRegexEmail;
 @RequestMapping("/app/users")
 @Slf4j
 @RequiredArgsConstructor
-/*
-  Controller란?
-  사용자의 Request를 전달받아 요청의 처리를 담당하는 Service, Prodiver 를 호출
- */
 public class UserController {
 
-//    final Logger logger = LoggerFactory.getLogger(this.getClass()); // Log를 남기기: 일단은 모르고 넘어가셔도 무방합니다.
-
-   // @Autowired  // 객체 생성을 스프링에서 자동으로 생성해주는 역할. 주입하려 하는 객체의 타입이 일치하는 객체를 자동으로 주입한다.
+    // @Autowired  // 객체 생성을 스프링에서 자동으로 생성해주는 역할. 주입하려 하는 객체의 타입이 일치하는 객체를 자동으로 주입한다.
     // IoC(Inversion of Control, 제어의 역전) / DI(Dependency Injection, 의존관계 주입)에 대한 공부하시면, 더 깊이 있게 Spring에 대한 공부를 하실 수 있을 겁니다!(일단은 모르고 넘어가셔도 무방합니다.)
     // IoC 간단설명,  메소드나 객체의 호출작업을 개발자가 결정하는 것이 아니라, 외부에서 결정되는 것을 의미
-    // DI 간단설명, 객체를 직접 생성하는 게 아니라 외부에서 생x성한 후 주입 시켜주는 방식
+    // DI 간단설명, 객체를 직접 생성하는 게 아니라 외부에서 생성한 후 주입 시켜주는 방식
+
     private final UserProvider userProvider;
     private final UserService userService;
     private final JwtService jwtService; // JWT부분은 7주차에 다루므로 모르셔도 됩니다!
 
-
-//    public UserController(UserProvider userProvider, UserService userService, JwtService jwtService) {
-//        this.userProvider = userProvider;
-//        this.userService = userService;
-//        this.jwtService = jwtService; // JWT부분은 7주차에 다루므로 모르셔도 됩니다!
-//    }
 
     // ******************************************************************************
 
@@ -61,7 +52,7 @@ public class UserController {
      */
     @ResponseBody
     @PostMapping("/sign-up")
-    public BaseResponse<List<GetUserRes>> createUser(@RequestBody PostUserReq postUserReq) {
+    public BaseResponse<PostUserRes> createUser(@RequestBody PostUserReq postUserReq) {
         if (postUserReq.getEmail() == null) {
             return new BaseResponse<>(POST_USERS_EMPTY_EMAIL);
         }
@@ -79,12 +70,8 @@ public class UserController {
             return new BaseResponse<>(POST_USERS_INVALID_EMAIL);
         }
         try {
-            Long userIdx= userService.createUser(postUserReq);
-            List<GetUserRes> getUserRes = new ArrayList<>();
-            List<PushNotificationAgreement> getAgreements = userProvider.getAgreements(userIdx);
-            User user = userProvider.getUser(userIdx);
-            getUserRes.add(new GetUserRes(user, new ArrayList<>(), getAgreements));
-            return new BaseResponse<>(getUserRes);
+            PostUserRes postUserRes = userService.createUser(postUserReq);
+            return new BaseResponse<>(postUserRes);
         } catch (BaseException exception) {
             return new BaseResponse<>((exception.getStatus()));
         }
@@ -94,8 +81,9 @@ public class UserController {
     @ResponseBody
     @PostMapping("/log-in")
     public BaseResponse<List<GetUserRes>> logIn(@RequestBody PostLoginReq postLoginReq) {
+
         try {
-            Long userIdx= userProvider.logIn(postLoginReq);
+            Long userIdx = userProvider.logIn(postLoginReq);
             User user = userProvider.getUser(userIdx);
             List<GetUserRes> getUserRes = new ArrayList<>();
             List<Address> getAddress = userProvider.getAddress(userIdx);
@@ -120,7 +108,7 @@ public class UserController {
             List<GetUserRes> getUserRes = new ArrayList<>(); // 반환할거 클라이언트에게
             List<User> getUsers = userProvider.getUsers();
             for (User getUser : getUsers) {
-                Long userIdx= getUser.getUserIdx();
+                Long userIdx = getUser.getUserIdx();
                 List<Address> getAddress = userProvider.getAddress(userIdx);
                 List<PushNotificationAgreement> getAgreements = userProvider.getAgreements(userIdx);
                 getUserRes.add(new GetUserRes(getUser, getAddress, getAgreements));
@@ -135,9 +123,15 @@ public class UserController {
     /**
      * 인덱스로 조회
      */
+    @SneakyThrows
     @ResponseBody
     @GetMapping("/{userIdx}")
     public BaseResponse<List<GetUserRes>> getUser(@PathVariable("userIdx") Long userIdx) {
+        Long userIdxByJwt = jwtService.getUserIdx();
+        //userIdx와 접근한 유저가 같은지 확인
+        if (!userIdx.equals(userIdxByJwt)) {
+            return new BaseResponse<>(INVALID_USER_JWT);
+        }
         try {
             List<GetUserRes> getUserRes = new ArrayList<>(); // 반환할거 클라이언트에게
             User getUser = userProvider.getUser(userIdx);
@@ -154,16 +148,27 @@ public class UserController {
     /**
      * 회원탈퇴
      */
+    @SneakyThrows
     @DeleteMapping("/{userIdx}")
     public BaseResponse<String> deleteUser(@PathVariable("userIdx") Long userIdx) {
+        Long userIdxByJwt = jwtService.getUserIdx();
+        //userIdx와 접근한 유저가 같은지 확인
+        if (!userIdx.equals(userIdxByJwt)) {
+            return new BaseResponse<>(INVALID_USER_JWT);
+        }
         try {
             Long id = userService.deleteUser(userIdx);
-            return new BaseResponse(true);
+            return new BaseResponse<>("true");
         } catch (BaseException exception) {
             return new BaseResponse<>((exception.getStatus()));
         }
     }
 
+    @GetMapping("/user-token/{userIdx}")
+    public BaseResponse<String> getUserToken(@PathVariable("userIdx") Long userIdx) {
+        String userJwtByIdx = jwtService.createJwt(userIdx);
+        return new BaseResponse<>(userJwtByIdx);
+    }
     /**
      * 유저정보변경 API
      * [PATCH] /users/:userIdx
@@ -192,8 +197,16 @@ public class UserController {
 //            return new BaseResponse<>((exception.getStatus()));
 //        }
 //    }
+    @SneakyThrows
     @PatchMapping("/{userIdx}/membership")
     public BaseResponse<String> updateMembership(@PathVariable("userIdx") Long userIdx, @RequestParam("memberType") String memberType) {
+
+        Long userIdxByJwt = jwtService.getUserIdx();
+        //userIdx와 접근한 유저가 같은지 확인
+        if (!userIdx.equals(userIdxByJwt)) {
+            return new BaseResponse<>(INVALID_USER_JWT);
+        }
+
         if (!memberType.equals("BASIC") && !memberType.equals("ROCKET")) {
             return new BaseResponse<>(PATCH_USERS_INVALID_MEMBERSHIP_LEVEL);
         }
@@ -205,8 +218,14 @@ public class UserController {
         }
     }
 
+    @SneakyThrows
     @PatchMapping("/{userIdx}/address")
     public BaseResponse<PatchAddressRes> updateAddress(@PathVariable Long userIdx, @RequestBody PatchAddressReq getAddressReq) {
+        Long userIdxByJwt = jwtService.getUserIdx();
+        //userIdx와 접근한 유저가 같은지 확인
+        if (!userIdx.equals(userIdxByJwt)) {
+            return new BaseResponse<>(INVALID_USER_JWT);
+        }
         try {
             if (getAddressReq.getIsDefault()) {    // 지금 입력하는 주소지가 기본 주소라면 이미있는 기본 주소지를 0으로 만듬
                 userService.initDefaultAddress(userIdx);
@@ -218,8 +237,14 @@ public class UserController {
         }
     }
 
+    @SneakyThrows
     @DeleteMapping("/{userIdx}/address")
     public BaseResponse<List<GetAddressRes>> deleteAddress(@PathVariable Long userIdx, @RequestParam("addressIdx") int addressIdx) {
+        Long userIdxByJwt = jwtService.getUserIdx();
+        //userIdx와 접근한 유저가 같은지 확인
+        if (!userIdx.equals(userIdxByJwt)) {
+            return new BaseResponse<>(INVALID_USER_JWT);
+        }
         try {
             userService.deleteAddress(addressIdx);
             return getListBaseResponse(userService.getAddress(userIdx), userIdx);
@@ -228,8 +253,14 @@ public class UserController {
         }
     }
 
+    @SneakyThrows
     @PostMapping("/{userIdx}/address")
     public BaseResponse<List<GetAddressRes>> insertAddress(@PathVariable Long userIdx, @RequestBody PostAddressReq postAddressReq) {
+        Long userIdxByJwt = jwtService.getUserIdx();
+        //userIdx와 접근한 유저가 같은지 확인
+        if (!userIdx.equals(userIdxByJwt)) {
+            return new BaseResponse<>(INVALID_USER_JWT);
+        }
         try {
             if (postAddressReq.getIsDefault()) {    // 지금 등록하는 주소지가 기본 주소라면 이미있는 기본 주소지를 0으로 만듬
                 userService.initDefaultAddress(userIdx);
@@ -245,8 +276,14 @@ public class UserController {
     }
 
 
+    @SneakyThrows
     @PatchMapping("{userIdx}/notification")
     public BaseResponse<GetPushNotificationAgreementRes> updatePushNotificationAgreement(@PathVariable Long userIdx, @RequestParam("notificationName") String notificationName) {
+        Long userIdxByJwt = jwtService.getUserIdx();
+        //userIdx와 접근한 유저가 같은지 확인
+        if (!userIdx.equals(userIdxByJwt)) {
+            return new BaseResponse<>(INVALID_USER_JWT);
+        }
         try {
             userIdx = userService.updatePushNotification(userIdx, notificationName);
             List<PushNotificationAgreement> pushNotificationAgreement = userProvider.getAgreements(userIdx);
@@ -258,6 +295,11 @@ public class UserController {
 
     @NotNull
     private BaseResponse<List<GetAddressRes>> getListBaseResponse(List<Address> address2, @PathVariable Long userIdx) throws BaseException {
+        Long userIdxByJwt = jwtService.getUserIdx();
+        //userIdx와 접근한 유저가 같은지 확인
+        if (!userIdx.equals(userIdxByJwt)) {
+            return new BaseResponse<>(INVALID_USER_JWT);
+        }
         List<GetAddressRes> getAddressRes = new ArrayList<>();
 
         for (Address address : address2) {
